UBUNTU: SAUCE: x86/virt/tdx: Remove the !KEXEC_CORE dependency
BugLink: https://bugs.launchpad.net/bugs/2121873

During kexec it is now guaranteed that all dirty cachelines of TDX
private memory are flushed before jumping to the new kernel.  The TDX
private memory from the old kernel will remain as TDX private memory in
the new kernel, but it is OK because kernel read/write to TDX private
memory will never cause machine check, except on the platforms with the
TDX partial write erratum, which has already been handled.

It is safe to allow kexec to work together with TDX now.  Remove the
!KEXEC_CORE dependency.

Signed-off-by: Kai Huang <kai.huang@intel.com>
Tested-by: Farrah Chen <farrah.chen@intel.com>
Reviewed-by: Rick Edgecombe <rick.p.edgecombe@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Link: https://lore.kernel.org/r/20250901160930.1785244-6-pbonzini@redhat.com
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>